+++ lib/operately/groups.ex
@@ -4,7 +4,7 @@ defmodule Operately.Groups do
   alias Operately.Repo
   alias Ecto.Multi
   alias Operately.Groups.{Group, Member, Contact}
-  alias Operately.Access.{Context, Filters}
+  alias Operately.Access.{Context, Fetch}
   alias Operately.People.Person
   alias Operately.Activities
 
@@ -50,7 +50,7 @@ defmodule Operately.Groups do
 
   def get_group_and_access_level(group_id, person_id) do
     from(g in Group, as: :resource, where: g.id == ^group_id)
-    |> Filters.get_resource_and_access_level(person_id)
+    |> Fetch.get_resource_and_access_level(person_id)
   end
 
   def get_group_by_name(name) do
